requirements.txt
python-chess
pandas
python-dotenv